+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bugfixes
 
 - [2816](https://github.com/influxdb/influxdb/pull/2816) -- enable UDP service. Thanks @renan-
+- [2824](https://github.com/influxdb/influxdb/pull/2824) -- Add missing call to WaitGroup.Done in execConn. Thanks @liyichao
 
 ## v0.9.0-rc32 [2015-06-07]