+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [2816](https://github.com/influxdb/influxdb/pull/2816) -- enable UDP service. Thanks @renan-
 - [2824](https://github.com/influxdb/influxdb/pull/2824) -- Add missing call to WaitGroup.Done in execConn. Thanks @liyichao
 - [2823](https://github.com/influxdb/influxdb/pull/2823) -- Convert OpenTSDB to a service.
+- [2838](https://github.com/influxdb/influxdb/pull/2838) -- Set auto-created retention policy period to infinite.
 
 ## v0.9.0-rc32 [2015-06-07]