+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - [#2857](https://github.com/influxdb/influxdb/issues/2857) -- Fix parsing commas in string field values.
 - [#2833](https://github.com/influxdb/influxdb/pull/2833) -- Make the default config valid.
 
+### Features
+- [2858](https://github.com/influxdb/influxdb/pull/2858) - Support setting openTSDB write consistency.
+
 ## v0.9.0-rc32 [2015-06-07]
 
 ### Release Notes